Move BranchSection sub-components into branch-section/ directory
- Relocate BranchCheckoutDialog / BranchCreateDialog / BranchDeleteDialog /
  BranchEntryList and branch-view-model(.test).ts under
  components/branch-section/, matching the commit-section convention
- Keep BranchSection.tsx (section entry point) directly under components/
- Update import paths only; no behavior change
