+++ apps/web/src/pages/SessionDetail/components/branch-section/BranchCheckoutDialog.tsx
