+++ apps/web/src/pages/SessionDetail/components/branch-section/BranchDeleteDialog.tsx
