+++ apps/web/src/pages/SessionDetail/components/branch-section/branch-view-model.test.ts
